Upgrading to play2.7 ...

+++ src/app/xcala.play/models/DataWithTotalCount.scala
@@ -1,3 +1,3 @@
 package xcala.play.models
 
-case class DataWithTotalCount[A](data: Seq[A], totalCount: Int)
+case class DataWithTotalCount[A](data: Seq[A], totalCount: Long)
